Tech Challenge Fase 2/requirements.txt
openai
python-dotenv
pandas
numpy
matplotlib
seaborn
imbalanced-learn
joblib
scikit-learn
fastapi
uvicorn
pydantic
azure-storage-blob
flask
python-multipart
imblearn